fix_emissive/requirements.txt
Pillow~=8.3.1